+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 1.3.1 (unreleased)
+## 1.3.2 (unreleased)
+
+## 1.3.1 (2025-08-03)
+
+- resource/porkbun_dnssec_record: Fix handling of deleted records when reading state.
+- resource/porkbun_url_forward: Fix handling of deleted records when reading state.
 
 ## 1.3.0 (2025-08-02)